📝💡✅ Huge update on test script process
Huge rework of script test.sh to be more userfriendly with dialog for
beginners and options to avoid dialog boxes.

Rework dockerfile for bash and zsh to not store main variables, i.e.
SHELL, SHELL_APP and PROMPT_VERSION in the image but are now passed as
argument determined by script test.sh

Update .envrc file in the image to not mess with the .envrc of the repo.
Now .envrc in the image press a line to ask user to press <Enter> once
in the docker to have all segments prompted.